create activity

+++ src/main/java/com/beton408/controller/FilesController.java
@@ -52,7 +52,7 @@ public class FilesController {
         user.setAvatar(randomFileName);
         // Lấy đường dẫn file cũ
 
-        if (oldFile != null) {
+        if (oldFile != null && oldFile != "User.png") {
             storageService.delete(oldFile);
         }
         String message = "";
